requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-mcp-adapters>=0.1.0
google-genai>=1.70.0
fastmcp>=2.0.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
httpx>=0.27.0
pillow>=10.0.0
python-dotenv>=1.0.0
